add: more javadocs and reorganized stuff a bit
