+++ src/main/java/lanat/helpRepresentation/descriptions/DescriptionFormatter.java
@@ -12,6 +12,12 @@ public final class DescriptionFormatter {
 
 	private DescriptionFormatter() {}
 
+	/**
+	 * Parses the description of the given user and replaces all tags with the content generated by them.
+	 * @param user the user whose description is being parsed
+	 * @param desc the description to parse
+	 * @return the parsed description
+	 */
 	public static @NotNull String parse(@NotNull NamedWithDescription user, @NotNull String desc) {
 		// if the description doesn't contain any tags, we can skip the parsing
 		if (!desc.contains(Character.toString(TAG_START)) && !desc.contains(Character.toString(TAG_END)))
@@ -53,6 +59,14 @@ public final class DescriptionFormatter {
 		return out.toString();
 	}
 
+	/**
+	 * Parses the description of the given user and replaces all tags with the content generated by them.
+	 * The description is taken from the given user by calling {@link NamedWithDescription#getDescription()}.
+	 * @param element the user whose description is being parsed
+	 * @return the parsed description, or null if the user has no description
+	 * @param <T> the type of the user
+	 * @see #parse(NamedWithDescription, String)
+	 */
 	public static <T extends CommandUser & NamedWithDescription>
 	@Nullable String parse(@NotNull T element) {
 		final var desc = element.getDescription();
@@ -62,6 +76,12 @@ public final class DescriptionFormatter {
 		return DescriptionFormatter.parse(element, desc);
 	}
 
+	/**
+	 * Parses the given tag and returns the content generated by its parser.
+	 * @param tagContents the contents of the tag, excluding the tag start and end characters
+	 * @param user the user whose description is being parsed
+	 * @return the content generated by the tag
+	 */
 	private static @NotNull String parseTag(@NotNull String tagContents, @NotNull NamedWithDescription user) {
 		if (tagContents.contains("=")) {
 			final var split = tagContents.split("=", 2);
