+++ src/main/java/lanat/helpRepresentation/descriptions/Tag.java
@@ -2,6 +2,10 @@ package lanat.helpRepresentation.descriptions;
 
 import lanat.NamedWithDescription;
 import lanat.helpRepresentation.descriptions.exceptions.UnknownTagException;
+import lanat.helpRepresentation.descriptions.tags.ColorTag;
+import lanat.helpRepresentation.descriptions.tags.DescTag;
+import lanat.helpRepresentation.descriptions.tags.FormatTag;
+import lanat.helpRepresentation.descriptions.tags.LinkTag;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
