+++ src/main/java/lanat/helpRepresentation/descriptions/tags/DescTag.java
@@ -1,11 +1,15 @@
-package lanat.helpRepresentation.descriptions;
+package lanat.helpRepresentation.descriptions.tags;
 
 import lanat.NamedWithDescription;
+import lanat.helpRepresentation.descriptions.DescriptionFormatter;
+import lanat.helpRepresentation.descriptions.RouteParser;
+import lanat.helpRepresentation.descriptions.Tag;
 import lanat.helpRepresentation.descriptions.exceptions.InvalidRouteException;
 import lanat.helpRepresentation.descriptions.exceptions.NoDescriptionDefinedException;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+/** Gets the description of the target object specified by the route. */
 public class DescTag extends Tag {
 	@Override
 	protected @NotNull String parse(@NotNull NamedWithDescription user, @Nullable String value) {
