+++ src/main/java/lanat/helpRepresentation/descriptions/tags/LinkTag.java
@@ -1,4 +1,4 @@
-package lanat.helpRepresentation.descriptions;
+package lanat.helpRepresentation.descriptions.tags;
 
 import lanat.Argument;
 import lanat.ArgumentGroup;
@@ -7,10 +7,13 @@ import lanat.NamedWithDescription;
 import lanat.helpRepresentation.ArgumentGroupRepr;
 import lanat.helpRepresentation.ArgumentRepr;
 import lanat.helpRepresentation.CommandRepr;
+import lanat.helpRepresentation.descriptions.RouteParser;
+import lanat.helpRepresentation.descriptions.Tag;
 import lanat.helpRepresentation.descriptions.exceptions.InvalidRouteException;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+/** Gets the representation of the target object specified by the route. */
 public class LinkTag extends Tag {
 	@Override
 	protected @NotNull String parse(@NotNull NamedWithDescription user, @Nullable String value) {
